pyproject.toml
[project]
name = "shitsu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.1.0",
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "isort>=8.0.0",
    "pillow>=12.1.1",
    "psycopg2>=2.9.11",
]
